restore.sh : handle increment with rdiff-backup method

+++ restore.sh
@@ -1,8 +1,24 @@
 #!/bin/bash
 
+# To restore an rdiff-backup increment:
+# 1) list increment with 
+#	rdiff-backup -l $REMOTE_HOST::$REMOTE_PATH
+#	ex :
+#	rdiff-backup -l backup@fr-hq-bkp-01::/data/backup-eqx-cas/rdiffbackup/fr-eqx-cas-09
+#		Found 4 increments:
+#		    increments.2016-07-16T17:55:23+02:00.dir   Sat Jul 16 17:55:23 2016       # <== let's restore this one
+#		    increments.2016-07-17T17:35:00+02:00.dir   Sun Jul 17 17:35:00 2016
+#		    increments.2016-07-18T16:32:02+02:00.dir   Mon Jul 18 16:32:02 2016
+#		    increments.2016-07-19T16:32:03+02:00.dir   Tue Jul 19 16:32:03 2016
+#		Current mirror: Wed Jul 20 17:31:03 2016
+#
+# 2) call the script without option -t and use -i with an increment date
+#  ./restore.sh -s fr-eqx-cas-09 -k 'campaign vasistas' -r backup@fr-hq-bkp-01 -f /data/backup-eqx-cas/rdiffbackup -m rdiff-backup -i "2016-07-16T17:55:23+02:00"
+
+
 PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
 
-while getopts "s:d:t:k:r:f:m:v:" opt; do
+while getopts "s:d:t:k:r:f:m:v:i:" opt; do
   case $opt in
     s)
       BACKUP_HOST=$OPTARG
@@ -28,22 +44,21 @@ while getopts "s:d:t:k:r:f:m:v:" opt; do
     v)
       VIGILANTE_ID=$OPTARG
       ;;
+    i)
+      INCREMENT_DATE=$OPTARG
+      ;;
     \?)
       echo "Invalid option: -$OPTARG" >&2
       ;;
   esac
 done
 
-if [ "$VIGILANTE_ID" = "" ]; then
-  echo "Missing required vigilante id (-v)"
-  exit
-fi
 if [ "$BACKUP_HOST" = "" ]; then
   echo "Missing required source host (-s)"
   exit
 fi
-if [ "$BACKUP_DATE" = "" -a "$TRIGGER_FILE" = "" ]; then
-  echo "Missing required date (-d) or trigger file (-t)"
+if [ "$BACKUP_DATE" = "" -a "$TRIGGER_FILE" = "" -a "$INCREMENT_DATE" = "" ]; then
+  echo "Missing required date (-d), trigger file (-t) or increment date (-i)"
   exit
 fi
 if [ "$DBS" = "" ]; then
@@ -77,7 +92,7 @@ if [[ ! `hostname` == *"casbkp"* ]]; then
   exit 1
 fi
 
-curl "http://vigilante.corp.withings.com/checkin/$VIGILANTE_ID?start" &> /dev/null
+[ -n "$VIGILANTE_ID"] && curl "http://vigilante.corp.withings.com/checkin/$VIGILANTE_ID?start" &> /dev/null
 
 # Real reload
 service cassandra stop
@@ -95,7 +110,7 @@ for tablefullpath in /var/lib/cassandra/data/$keyspacename/*; do
       if [ $METHOD = "rsync" ]; then
         BACKUP_FULLPATH="data/"$BACKUP_DATE"/"$keyspacename"/"$table"*"
       elif [ $METHOD = "rdiff-backup" ]; then
-	BACKUP_FULLPATH=$keyspacename"/"$table"*"
+	BACKUP_FULLPATH=$( ssh $REMOTE_HOST "cd $REMOTE_PATH/$BACKUP_HOST; ls -d $keyspacename/$table*" )
       else
         BACKUP_FULLPATH=$BACKUP_DATE"/"$keyspacename"/"$table"*.tbz2"
       fi
@@ -104,8 +119,12 @@ for tablefullpath in /var/lib/cassandra/data/$keyspacename/*; do
         echo "table:[$table] method:[$METHOD] "$BACKUP_FULLPATH" TO "$tablefullpath
 	MESSAGE+="Restoring $keyspacename:${table%-*}"$'\n'
         find "$tablefullpath/" -type f -delete
-        if [ $METHOD = "rsync" ] || [ $METHOD = "rdiff-backup" ]; then
+        if [ $METHOD = "rsync" ] || [ $METHOD = "rdiff-backup" -a -z "$INCREMENT_DATE" ]; then
           CMD="rsync -a --delete $REMOTE_HOST:$REMOTE_PATH/$BACKUP_HOST/$BACKUP_FULLPATH/ $tablefullpath/"
+          echo "  "$CMD
+          $CMD
+        elif [ $METHOD = "rdiff-backup" ] && [ -n "$INCREMENT_DATE" ]; then # restore increment
+          CMD="rdiff-backup --force -r $INCREMENT_DATE $REMOTE_HOST::$REMOTE_PATH/$BACKUP_HOST/$BACKUP_FULLPATH/ $tablefullpath/"
           echo "  "$CMD
           $CMD
         else
@@ -147,9 +166,13 @@ for keyspacename in $DBS; do
     tablefullpath=/var/lib/cassandra/data/$keyspacename/$cf_name
     mkdir -p $tablefullpath
     MESSAGE+="Restoring $keyspacename:$cf_name"$'\n'
-    if [ $METHOD = "rsync" ] || [ $METHOD = "rdiff-backup" ]; then
+    if [ $METHOD = "rsync" ] || [ $METHOD = "rdiff-backup" -a -z "$INCREMENT_DATE" ]; then
       echo "Copy $ksremotefullpath/$cf into $tablefullpath"
       CMD="rsync -a --delete $REMOTE_HOST:$ksremotefullpath/$cf/ $tablefullpath/"
+      echo "  "$CMD
+      $CMD
+    elif [ $METHOD = "rdiff-backup" ] && [ -n "$INCREMENT_DATE" ]; then # restore increment
+      CMD="rdiff-backup --force -r $INCREMENT_DATE $REMOTE_HOST::$ksremotefullpath/$cf/ $tablefullpath/"
       echo "  "$CMD
       $CMD
     else
@@ -176,5 +199,5 @@ echo "All done"
 #STATUS=$?
 STATUS=0
 
-curl --data "status=$STATUS&message=$MESSAGE" http://vigilante.corp.withings.com/checkin/$VIGILANTE_ID &> /dev/null
+[ -n "$VIGILANTE_ID"] && curl --data "status=$STATUS&message=$MESSAGE" http://vigilante.corp.withings.com/checkin/$VIGILANTE_ID &> /dev/null
